requirements.txt
numpy==1.24.2
customtkinter==5.1.3
Pillow==9.3.0
face_recognition==1.3.0
opencv-python==4.7.0.72
threading==0.1.0